Finished the integration of the wheel as a component (not a separate window), and added the functionality to select an arbitrary topic.
