+++ webapp/src/css/Roulette.css
@@ -1,17 +1,18 @@
 /* Roulette.css */
 .roulette-container {
+    position: relative;
     display: flex;
     flex-direction: column;
     align-items: center;
     gap: 3rem;
-    padding: 2rem;
+    padding: 1rem;
 }
 
 .roulette-wheel {
     position: relative;
-    width: 30rem;
-    height: 30rem;
-    border: 5px solid;
+    width: 28rem;
+    height: 28rem;
+    border: 3px solid;
     border-radius: 50%;
     display: flex;
     align-items: center;
@@ -27,7 +28,7 @@
 
 .pointer {
     position: absolute;
-    top: 45%;
+    top: 28rem;
     left: 50%;
     transform: translateX(-50%);
     width: 0;
